+++ backend_ecoride/src/Controller/DriveController.php
@@ -4,21 +4,24 @@ namespace App\Controller;
 
 use App\DTO\Drive\{DriveDTO, DriveSearchDTO};
 
-use App\Service\Drive\{
+use App\Service\Drive\Manage\{
     CreateDriveService,
-    ReadDriveService,
     UpdateDriveService,
     ListDrivePaginatedService,
-    JoinDriveService,
-    LeaveDriveService,
+
     StartDriveService,
     FinishDriveService,
     CancelDriveService,
-    SearchDriveService,
+};
 
-    ConfirmParticipationService,
-    DisputeParticipationService,
-    ResolveDisputeService,
+use App\Service\Drive\Query\{
+    ReadDriveService,
+    SearchDriveService,
+};
+
+use App\Service\Drive\Participation\{
+    JoinDriveService,
+    LeaveDriveService,
 };
 
 use App\Service\Access\AccessControlService;
@@ -464,187 +467,6 @@ final class DriveController extends AbstractController
             return new JsonResponse(
                 ['error' => $e->getMessage()],
                 JsonResponse::HTTP_INTERNAL_SERVER_ERROR
-            );
-        }
-    }
-
-    #[Route('/{identifier}/confirm', name: 'confirm_participation', requirements: ['identifier' => '.+'], methods: 'POST')]
-    public function confirmParticipation(
-        string $identifier,
-        ConfirmParticipationService $confirmService,
-    ): JsonResponse {
-        try {
-            $this->accessControl->denyUnlessLogged();
-            $this->accessControl->denyIfBanned();
-
-            $confirmService->confirm($identifier);
-
-            return new JsonResponse(
-                data: ['message' => 'Participation confirmed'],
-                status: JsonResponse::HTTP_OK
-            );
-
-        } catch (AccessDeniedHttpException $e) {
-            return new JsonResponse(
-                ['error' => $e->getMessage()],
-                JsonResponse::HTTP_FORBIDDEN
-            );
-        } catch (NotFoundHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_NOT_FOUND
-            );
-        } catch (BadRequestHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_BAD_REQUEST
-            );
-        } catch (\Exception $e) {
-            return new JsonResponse(
-                data: ['error' => "An internal server error as occured"],
-                status: JsonResponse::HTTP_INTERNAL_SERVER_ERROR
-            );
-        }
-    }
-
-    #[Route('/{identifier}/dispute', name: 'dispute_participation', requirements: ['identifier' => '.+'], methods: 'POST')]
-    public function dispute(
-        string $identifier,
-        Request $request,
-        DisputeParticipationService $disputeService
-    ): JsonResponse {
-        try {
-            $this->accessControl->denyUnlessLogged();
-            $this->accessControl->denyIfBanned();
-
-            $raw = $request->getContent();
-            if ($raw === null || $raw === '') {
-                $payload = [];
-            } else {
-                $payload = json_decode($raw, true);
-                if ($payload === null && json_last_error() !== JSON_ERROR_NONE) {
-                    throw new BadRequestHttpException("Invalid JSON format");
-                }
-                if (!is_array($payload)) {
-                    throw new BadRequestHttpException("Invalid JSON format");
-                }
-            }
-
-            $comment = null;
-            if (array_key_exists('comment', $payload)) {
-                if (!is_string($payload['comment'])) {
-                    throw new BadRequestHttpException("Comment must be string");
-                }
-                $comment = trim($payload['comment']);
-                if ($comment === '') {
-                    $comment = null;
-                }
-            }
-
-            $disputeService->dispute($identifier, $comment);
-
-            return new JsonResponse(
-                data: ['message' => 'Dispute opened'],
-                status: JsonResponse::HTTP_OK
-            );
-
-        } catch (AccessDeniedHttpException $e) {
-            return new JsonResponse(
-                ['error' => $e->getMessage()],
-                JsonResponse::HTTP_FORBIDDEN
-            );
-        } catch (NotFoundHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_NOT_FOUND
-            );
-        } catch (BadRequestHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_BAD_REQUEST
-            );
-        } catch (\Exception $e) {
-            return new JsonResponse(
-                data: ['error' => "An internal server error as occured"],
-                status: JsonResponse::HTTP_INTERNAL_SERVER_ERROR
-            );
-        }
-    }
-
-    #[Route('/{identifier}/dispute/{participantUuid}/resolve', name: 'resolve_dispute', requirements: ['identifier' => '.+'], methods: 'POST')]
-    public function resolve(
-        string $identifier,
-        string $participantUuid,
-        Request $request,
-        ResolveDisputeService $resolveDisputeService,
-    ): JsonResponse {
-        try {
-            $this->accessControl->denyUnlessLogged();
-            $this->accessControl->denyIfBanned();
-            $this->accessControl->denyUnlessEmployee();
-
-            $raw = $request->getContent();
-            if ($raw === null || $raw === '') {
-                $payload = [];
-            } else {
-                $payload = json_decode($raw, true);
-                if ($payload === null && json_last_error() !== JSON_ERROR_NONE) {
-                    throw new BadRequestHttpException("Invalid JSON format");
-                }
-                if (!is_array($payload)) {
-                    throw new BadRequestHttpException("Invalid JSON format");
-                }
-            }
-
-            if (!array_key_exists('action', $payload)) {
-                throw new BadRequestHttpException("Field 'action' is required");
-            }
-            if (!is_string($payload['action'])) {
-                throw new BadRequestHttpException("Field 'action' must be string");
-            }
-
-            $action = trim($payload['action']);
-            if ($action === '') {
-                throw new BadRequestHttpException("Field 'action' cannot be empty");
-            }
-
-            $comment = null;
-            if (array_key_exists('comment', $payload)) {
-                if (!is_string($payload['comment'])) {
-                    throw new BadRequestHttpException("Comment must be string");
-                }
-                $comment = trim($payload['comment']);
-                if ($comment === '') {
-                    $comment = null;
-                }
-            }
-
-            $resolveDisputeService->resolve($identifier, $participantUuid, $action, $comment);
-
-            return new JsonResponse(
-                ['message' => 'Dispute resolved'],
-                status: JsonResponse::HTTP_OK
-            );
-
-        } catch (AccessDeniedHttpException $e) {
-            return new JsonResponse(
-                ['error' => $e->getMessage()],
-                JsonResponse::HTTP_FORBIDDEN
-            );
-        } catch (NotFoundHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_NOT_FOUND
-            );
-        } catch (BadRequestHttpException $e) {
-            return new JsonResponse(
-                data: ['error' => $e->getMessage()],
-                status: JsonResponse::HTTP_BAD_REQUEST
-            );
-        } catch (\Exception $e) {
-            return new JsonResponse(
-                data: ['error' => "An internal server error as occured"],
-                status: JsonResponse::HTTP_INTERNAL_SERVER_ERROR
             );
         }
     }
