+++ backend_ecoride/src/Service/Drive/Manage/FinishDriveService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Drive;
+namespace App\Service\Drive\Manage;
 
 use App\Entity\Drive;
 use App\Repository\DriveRepository;
