+++ backend_ecoride/src/Service/Drive/Manage/ListDrivePaginatedService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service\Drive;
+namespace App\Service\Drive\Manage;
 
 use App\Entity\User;
